requirements.txt
django
djangorestframework
django-autoslug
django-celery
django-compressor
django-cors-headers
django-filter
django-jsonfield
django-model-utils
django-picklefield
factory_boy
mock
pillow
python-dateutil
python-memcached
python-money

git+ssh://git@bitbucket.org/litedesk/cross7-lib-active_directory
git+ssh://git@bitbucket.org/litedesk/cross7-lib-airwatch
